add slice for regular index
