+++ packages/compass-indexes/src/components/indexes/indexes.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { cleanup, render, screen, within } from '@testing-library/react';
 import { expect } from 'chai';
 import AppRegistry from 'hadron-app-registry';
-import type { IndexDefinition } from '../../modules/indexes';
+import type { IndexDefinition } from '../../modules/regular-indexes';
 import { Indexes } from './indexes';
 
 const renderIndexes = (
