+++ packages/compass-indexes/src/modules/index.ts
@@ -15,13 +15,9 @@ import isReadonlyView, {
 import description, {
   INITIAL_STATE as DESCRIPTION_INITIAL_STATE,
 } from './description';
-import indexes, { INITIAL_STATE as INDEXES_INITIAL_STATE } from './indexes';
-import sortOrder, {
-  INITIAL_STATE as SORT_ORDER_INITIAL_STATE,
-} from './sort-order';
-import sortColumn, {
-  INITIAL_STATE as SORT_COLUMN_INITIAL_STATE,
-} from './sort-column';
+import regularIndexes, {
+  INITIAL_STATE as REGULAR_INDEXES_INITIAL_STATE,
+} from './regular-indexes';
 import error, { INITIAL_STATE as ERROR_INITIAL_STATE } from './error';
 import serverVersion, {
   INITIAL_STATE as SV_INITIAL_STATE,
@@ -29,26 +25,18 @@ import serverVersion, {
 import namespace, {
   INITIAL_STATE as NAMESPACE_INITIAL_STATE,
 } from './namespace';
-import isRefreshing, {
-  INITIAL_STATE as REFRESHING_INITIAL_STATE,
-} from './is-refreshing';
-import inProgressIndexes from './in-progress-indexes';
 
 const reducer = combineReducers({
-  indexes,
   isWritable,
   isReadonly,
   isReadonlyView,
   description,
   appRegistry,
   dataService,
-  sortOrder,
-  sortColumn,
   error,
   serverVersion,
   namespace,
-  isRefreshing,
-  inProgressIndexes,
+  regularIndexes,
 });
 
 export type RootState = ReturnType<typeof reducer>;
@@ -61,13 +49,10 @@ const rootReducer = (state: RootState, action: AnyAction): RootState => {
       isReadonly: READONLY_INITIAL_STATE,
       isReadonlyView: READONLY_VIEW_INITIAL_STATE,
       description: DESCRIPTION_INITIAL_STATE,
-      indexes: INDEXES_INITIAL_STATE,
       serverVersion: SV_INITIAL_STATE,
-      sortOrder: SORT_ORDER_INITIAL_STATE,
-      sortColumn: SORT_COLUMN_INITIAL_STATE,
       error: ERROR_INITIAL_STATE,
       namespace: NAMESPACE_INITIAL_STATE,
-      isRefreshing: REFRESHING_INITIAL_STATE,
+      regularIndexes: REGULAR_INDEXES_INITIAL_STATE,
     };
   }
   return reducer(state, action);
