+++ packages/compass-indexes/src/utils/is-action.ts
@@ -1,0 +1,8 @@
+import type { AnyAction } from 'redux';
+
+export function isAction<A extends AnyAction>(
+  action: AnyAction,
+  type: A['type']
+): action is A {
+  return action.type === type;
+}
