+++ packages/compass-indexes/test/fixtures/regular-indexes.ts
@@ -1,0 +1,364 @@
+export const indexesList = [
+  {
+    name: '_id_',
+    version: 2,
+    extra: {},
+    key: { _id: 1 },
+    usageHost: 'computername.local:27017',
+    usageCount: 6,
+    usageSince: new Date('2019-02-08T10:21:49.176Z'),
+    size: 135168,
+  },
+  {
+    name: 'CCCC',
+    version: 2,
+    extra: {},
+    key: { cccc0: -1, cccc1: '2dsphere', cccc2: 1 },
+    '2dsphereIndexVersion': 3,
+    usageHost: 'computername.local:27017',
+    usageCount: 5,
+    usageSince: new Date('2019-02-08T14:38:56.147Z'),
+    size: 4096,
+  },
+  {
+    name: 'AAAA',
+    version: 2,
+    extra: {},
+    key: { aaaa: -1 },
+    expireAfterSeconds: 300,
+    partialFilterExpression: { y: 1 },
+    usageHost: 'computername.local:27017',
+    usageCount: 4,
+    usageSince: new Date('2019-02-08T14:39:56.285Z'),
+    size: 4096,
+  },
+  {
+    name: 'BBBB',
+    version: 2,
+    extra: {},
+    key: { 'bbbb.abcd': 1 },
+    collation: {
+      locale: 'ar',
+      caseLevel: true,
+      caseFirst: 'lower',
+      strength: 3,
+      numericOrdering: false,
+      alternate: 'non-ignorable',
+      maxVariable: 'space',
+      normalization: true,
+      backwards: true,
+      version: '57.1',
+    },
+    usageHost: 'computername.local:27017',
+    usageCount: 10,
+    usageSince: new Date('2019-02-08T14:40:13.609Z'),
+    size: 94208,
+  },
+];
+
+export const defaultSortedIndexes = [
+  {
+    ns: 'citibike.trips',
+    key: { aaaa: -1 },
+    name: 'AAAA',
+    size: 4096,
+    usageCount: 4,
+    usageSince: new Date('2019-02-08T14:39:56.285Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {
+      expireAfterSeconds: 300,
+      partialFilterExpression: { y: 1 },
+    },
+    id: 'citibike.trips.AAAA',
+    unique: false,
+    sparse: false,
+    ttl: true,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: true,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['partial', 'ttl'],
+  },
+  {
+    ns: 'citibike.trips',
+    key: { 'bbbb.abcd': 1 },
+    name: 'BBBB',
+    size: 94208,
+    usageCount: 10,
+    usageSince: new Date('2019-02-08T14:40:13.609Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {
+      collation: {
+        locale: 'ar',
+        caseLevel: true,
+        caseFirst: 'lower',
+        strength: 3,
+        numericOrdering: false,
+        alternate: 'non-ignorable',
+        maxVariable: 'space',
+        normalization: true,
+        backwards: true,
+        version: '57.1',
+      },
+    },
+    id: 'citibike.trips.BBBB',
+    unique: false,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: true,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['collation'],
+  },
+  {
+    ns: 'citibike.trips',
+    key: { cccc0: -1, cccc1: '2dsphere', cccc2: 1 },
+    name: 'CCCC',
+    size: 4096,
+    usageCount: 5,
+    usageSince: new Date('2019-02-08T14:38:56.147Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: { '2dsphereIndexVersion': 3 },
+    id: 'citibike.trips.CCCC',
+    unique: false,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: true,
+    compound: true,
+    single: false,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'geospatial',
+    cardinality: 'compound',
+    properties: [],
+  },
+  {
+    ns: 'citibike.trips',
+    key: { _id: 1 },
+    name: '_id_',
+    size: 135168,
+    usageCount: 6,
+    usageSince: new Date('2019-02-08T10:21:49.176Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {},
+    id: 'citibike.trips._id_',
+    unique: true,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['unique'],
+  },
+];
+
+export const usageSortedIndexes = [
+  {
+    name: 'AAAA',
+    usageCount: 4,
+
+    ns: 'citibike.trips',
+    key: { aaaa: -1 },
+    size: 4096,
+    usageSince: new Date('2019-02-08T14:39:56.285Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {
+      expireAfterSeconds: 300,
+      partialFilterExpression: { y: 1 },
+    },
+    id: 'citibike.trips.AAAA',
+    unique: false,
+    sparse: false,
+    ttl: true,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: true,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['partial', 'ttl'],
+  },
+  {
+    name: 'CCCC',
+    usageCount: 5,
+
+    ns: 'citibike.trips',
+    key: { cccc0: -1, cccc1: '2dsphere', cccc2: 1 },
+    size: 4096,
+    usageSince: new Date('2019-02-08T14:38:56.147Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: { '2dsphereIndexVersion': 3 },
+    id: 'citibike.trips.CCCC',
+    unique: false,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: true,
+    compound: true,
+    single: false,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'geospatial',
+    cardinality: 'compound',
+    properties: [],
+  },
+  {
+    name: '_id_',
+    usageCount: 6,
+
+    ns: 'citibike.trips',
+    key: { _id: 1 },
+    size: 135168,
+    usageSince: new Date('2019-02-08T10:21:49.176Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {},
+    id: 'citibike.trips._id_',
+    unique: true,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: false,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['unique'],
+  },
+  {
+    name: 'BBBB',
+    usageCount: 10,
+
+    ns: 'citibike.trips',
+    key: { 'bbbb.abcd': 1 },
+    size: 94208,
+    usageSince: new Date('2019-02-08T14:40:13.609Z'),
+    usageHost: 'computername.local:27017',
+    version: 2,
+    extra: {
+      collation: {
+        locale: 'ar',
+        caseLevel: true,
+        caseFirst: 'lower',
+        strength: 3,
+        numericOrdering: false,
+        alternate: 'non-ignorable',
+        maxVariable: 'space',
+        normalization: true,
+        backwards: true,
+        version: '57.1',
+      },
+    },
+    id: 'citibike.trips.BBBB',
+    unique: false,
+    sparse: false,
+    ttl: false,
+    hashed: false,
+    geo: false,
+    compound: false,
+    single: true,
+    partial: false,
+    text: false,
+    wildcard: false,
+    collation: true,
+    clustered: false,
+    columnstore: false,
+    type: 'regular',
+    cardinality: 'single',
+    properties: ['collation'],
+  },
+];
+
+export const inProgressIndexes = [
+  {
+    name: 'AAAA',
+    version: 2,
+    extra: {
+      status: 'inprogress',
+    },
+    key: {
+      aaaa: -1,
+    },
+    expireAfterSeconds: 300,
+    partialFilterExpression: {
+      y: 1,
+    },
+    usageHost: 'computername.local:27017',
+    usageCount: 4,
+    usageSince: new Date('2019-02-08T14:39:56.285Z'),
+    size: 4096,
+  },
+  {
+    id: 'citibike.trips.z',
+    extra: {
+      status: 'inprogress',
+    },
+    key: {
+      z: 1,
+    },
+    fields: [
+      {
+        field: 'z',
+        value: 1,
+      },
+    ],
+    name: 'z',
+    ns: 'citibike.trips',
+    size: 0,
+    relativeSize: 0,
+    usageCount: 0,
+  },
+];
